changes to measureOperationsUtils
